backend/requirements.txt
Django==2.2.18
django-cors-middleware==1.5.0
django-crispy-forms==1.9.2
django-filter==2.2.0
django-guardian==1.4.9
django-model-utils==3.2.0
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
django-celery-results==1.0.1
djoser==2.0.3
dj-database-url==0.5.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
six==1.12.0
whitenoise==5.0.1
celery==4.2.1
python-dateutil==2.8.1
pyfcm==1.4.7
drf-yasg==1.17.1
